Use DEFAULT_MAX_PERMITS constant in getSharedArenaMaxPermitsSysprop (#15908) * adjust SharedArenaMaxPermits to RefCountedSharedArena.DEFAULT_MAX_PERMITS
* add changes

* move change entry to 10.5

* adjust factory class to have access to default constant
