+++ lucene/core/src/java/org/apache/lucene/store/MMapDirectory.java
@@ -29,7 +29,6 @@ import java.util.concurrent.Future;
 import java.util.function.BiFunction;
 import java.util.function.BiPredicate;
 import java.util.function.Function;
-import java.util.logging.Logger;
 import org.apache.lucene.index.IndexFileNames;
 import org.apache.lucene.util.Constants;
 
@@ -395,33 +394,15 @@ public class MMapDirectory extends FSDirectory {
     }
   }
 
-  private static int getSharedArenaMaxPermitsSysprop() {
-    int ret = 1024; // default value
-    try {
-      String str = System.getProperty(SHARED_ARENA_MAX_PERMITS_SYSPROP);
-      if (str != null) {
-        ret = Integer.parseInt(str);
-      }
-    } catch (@SuppressWarnings("unused") NumberFormatException | SecurityException ignored) {
-      Logger.getLogger(MMapDirectory.class.getName())
-          .warning(
-              "Cannot read sysprop "
-                  + SHARED_ARENA_MAX_PERMITS_SYSPROP
-                  + ", so the default value will be used.");
-    }
-    return ret;
-  }
-
   private static <A> MMapIndexInputProvider<A> lookupProvider() {
-    final var maxPermits = getSharedArenaMaxPermitsSysprop();
     final var lookup = MethodHandles.lookup();
     try {
       final var cls = lookup.findClass("org.apache.lucene.store.MemorySegmentIndexInputProvider");
       // we use method handles, so we do not need to deal with setAccessible as we have private
       // access through the lookup:
-      final var constr = lookup.findConstructor(cls, MethodType.methodType(void.class, int.class));
+      final var constr = lookup.findConstructor(cls, MethodType.methodType(void.class));
       try {
-        return (MMapIndexInputProvider<A>) constr.invoke(maxPermits);
+        return (MMapIndexInputProvider<A>) constr.invoke();
       } catch (RuntimeException | Error e) {
         throw e;
       } catch (Throwable th) {
